requirements.txt
requests
openai
google-genai
redis
python-dotenv
Pillow
